fix e2e-tests/examples + feature freeze
this commit fixes the open source examples + e2e tests. all we need
to do is get them running on every commit, and then we can publish
this release.

this brings us up to feature compatibility with the internal version
of coworker this library started from. metrics, and configuration
of everything.

the first open sourcing of this was good (but had some bugs!), we've
also kept some parts of the interface "unstable" in this public version.
However, at this point in time we're happy to announce that the
interface will be stabilized as of 1.2.0. Not to mention internal
clients using coworker, will start pointing to the OSS version
of coworker.

this is pretty big news, and we'll be happy to finally deliver quality.